XW-3310 | ownerInjection in rest of models
