+++ app/controllers/article-edit.js
@@ -1,5 +1,4 @@
 import Ember from 'ember';
-import ArticleEditModel from '../models/article-edit';
 import {track, trackActions} from '../utils/track';
 import {normalizeToUnderscore} from '../utils/string';
