+++ app/models/notifications/contributor.js
@@ -28,15 +28,16 @@ DiscussionContributor.reopenClass({
 		});
 	},
 	/**
+	 * @param {*} ownerInjection
 	 * @param {object} data
 	 *
 	 * @returns {Ember.Object}
 	 */
-	create(data) {
+	create(ownerInjection, data) {
 		let result = null;
 
 		if (data) {
-			result = this._super({
+			result = this._super(ownerInjection, {
 				avatarUrl: data.avatarUrl,
 				badgePermission: data.badgePermission,
 				id: data.id,
